requirements.txt
atomicwrites==1.4.1
attrs==23.1.0
colorama==0.4.6
Django==2.2.16
django-debug-toolbar==3.2.4
exceptiongroup==1.1.1
flake8==6.0.0
flake8-isort==6.0.0
gitdb==4.0.10
GitPython==3.1.31
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==6.2.5
pytest-django==4.5.2
pytest-env==0.8.2
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
